+++ lib/async.dart
@@ -4,5 +4,6 @@
 
 library dart.pkg.async;
 
+export "src/future_group.dart";
 export "stream_zip.dart";
 export "result.dart";